Fix Quizzes page
