+++ src/App.js
@@ -9,7 +9,7 @@ import ResetPassword from 'pages/ResetPassword';
 import Footer from 'components/Footer';
 import Error from 'pages/Error';
 import Home from 'pages/Home/Home';
-import QuizContent from 'pages/QuizUi/QuizContent';
+import QuizContent from 'pages/Quiz/QuizContent';
 import LogoutModal from 'components/LogoutModal';
 import highlight_1 from './assets/images/highlight_effective_knowledge_testing.svg';
 import highlight_2 from './assets/images/highlight_preparing_for_a_job_interview.svg';
@@ -19,8 +19,8 @@ import CustomizedSnackbars from 'components/Snackbar';
 import Header from './components/Header/Header';
 import SettingsRoundedIcon from '@mui/icons-material/Settings';
 import LogoutRoundedIcon from '@mui/icons-material/LogoutRounded';
-import { Quizzes } from './pages/Main';
-import { Favorites } from './pages/Main';
+import { Quizzes } from './pages/Main/Quizzes';
+import { Favorites } from './pages/Main/Favorites';
 import About from './pages/About/About';
 import NavBar from './components/NavBar';
 import Notes from './pages/Notes';
