+++ src/config.js
@@ -2,7 +2,7 @@ export let BASE_URL;
 
 if (process.env.NODE_ENV === 'production') {
   // Set production API URL
-  BASE_URL = 'https://jsquiz-u1ri.onrender.com';
+  BASE_URL = 'https://jsquiz-u1ri.onrender.com/api/v1';
 } else {
   // Set local development API URL
   BASE_URL = 'http://localhost:8000/api/v1';
